+++ src/components/pages/register/add-photos/add-photos.js
@@ -18,7 +18,6 @@ const AddPhotos = () => {
         <UploadImages
           func={handleOnChange}
           name="main-image"
-        //   defaultValue={[]}
         />
       </div>
     </div>